pyproject.toml
[tool.poetry]
name = "struzstartup"
version = "0.1.0"
description = "Project description."
authors = ["My Name <me@myself.com>"]

[tool.poetry.dependencies]
python = "^3.7"
apispec = "*"
aws-xray-sdk = "*"
boto3 = "*"
factory-boy = "*"
faker = "*"
flask = "*"
flask-cors = "*"
flask-crud = {git = "https://github.com/jetbridge/flask-crud.git"}
flask-jwt-extended = "*"
flask-migrate = "*"
flask-script = "*"
flask-smorest = "*"
flask-sqlalchemy = "*"
jetkit-flask = "==9.0.1"
marshmallow = "*"
marshmallow-enum = "*"
marshmallow-sqlalchemy = "*"
nplusone = "*"
psycopg2-binary = "*"
requests = "*"
sqlalchemy = "*"
stripe = "^2.45.0"
pre-commit-hooks = "^2.5.0"
request = "^2019.4.13"
email_normalize = "^0.2.1"

[tool.poetry.dev-dependencies]
bento-cli = "*"
black = "^19.3b0"
cfn-lint = "*"
flake8 = "*"
flask-shell-ipython = "*"
ipdb = "*"
mypy = "*"
pre-commit = "*"
pytest = "*"
pytest-factoryboy = "*"
pytest-flask-sqlalchemy = "*"
pytest-postgresql = "*"
pytest-xdist = "*"
flake8-docstrings = "^1.5.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
